PC-53 feat: padroniza Sidebar via layout
- Criado layout.tsx em /auth para uso compartilhado do Sidebar
- Adicionado componente Sidebar.tsx em /components/ui
- Removido Sidebar individuais das páginas de animal, fazenda e dashboard
