+++ src/app/auth/fazenda/cadastrar/page.tsx
@@ -31,7 +31,6 @@ const sizeLabel = (s: 1 | 2 | 3) => (s === 1 ? "Pequena" : s === 2 ? "Média" : 
 export default function CadastrarFazendaPage() {
   const { theme } = useTheme();
   const [mounted, setMounted] = useState(false);
-  const [sidebarOpen, setSidebarOpen] = useState(true);
 
   const [estados, setEstados] = useState<Estado[]>([]);
   const [municipios, setMunicipios] = useState<Municipio[]>([]);
@@ -48,7 +47,6 @@ export default function CadastrarFazendaPage() {
 
   useEffect(() => setMounted(true), []);
   const darkMode = theme === "dark";
-  const toggleSidebar = () => setSidebarOpen(!sidebarOpen);
 
   // Carregar UFs ao montar
   useEffect(() => {
@@ -84,89 +82,6 @@ export default function CadastrarFazendaPage() {
 
   return (
     <div className="flex min-h-screen transition-colors duration-500 bg-background text-foreground">
-      {/* Sidebar */}
-      <aside
-        className={`flex flex-col justify-between shadow-lg transition-all duration-300 bg-sidebar text-sidebar-foreground ${sidebarOpen ? "w-64 p-6" : "w-20 p-2"
-          }`}
-      >
-        <div>
-          <div className="flex items-center justify-between mb-10">
-            <Image
-              src="/images/cowuai-logo.png"
-              alt="CowUai Logo"
-              width={sidebarOpen ? 100 : 40}
-              height={sidebarOpen ? 100 : 40}
-            />
-            <PiKeyReturn
-              size={30}
-              className="text-red-600 cursor-pointer"
-              onClick={toggleSidebar}
-              title={sidebarOpen ? "Recolher" : "Expandir"}
-            />
-          </div>
-
-          {/* Navegação */}
-          <nav className="space-y-2">
-            <Link
-              href="/dashboard"
-              className="flex items-center gap-2 px-4 py-2 rounded-lg font-medium transition-colors duration-500 bg-sidebar text-sidebar-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground"
-            >
-              <LuLayoutDashboard size={20} className="text-red-600" />
-              {sidebarOpen && <span>Dashboard</span>}
-            </Link>
-
-            <details className="group">
-              <summary className="flex items-center gap-2 cursor-pointer px-4 py-2 rounded-lg font-medium transition-colors duration-500 bg-sidebar text-sidebar-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                <PiCow size={20} className="text-red-600" />
-                {sidebarOpen && <span>Gerenciar Animais</span>}
-                <FaChevronRight className="transition-transform duration-300 group-open:rotate-90 text-red-600" />
-              </summary>
-
-              {sidebarOpen && (
-                <ul className="mt-2 space-y-1 text-md bg-sidebar text-sidebar-foreground rounded-md p-2">
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="/auth/animal/cadastrar">Cadastrar Animal</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="#">Atualizar Animal</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="#">Excluir Animal</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="#">Visualizar Animal</Link>
-                  </li>
-                </ul>
-              )}
-            </details>
-
-            <details className="group mt-2" open>
-              <summary className="flex items-center gap-2 cursor-pointer px-4 py-2 rounded-lg font-medium transition-colors duration-500 bg-sidebar text-sidebar-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                <PiUsersThree size={20} className="text-red-600" />
-                {sidebarOpen && <span>Gerenciar Fazendas</span>}
-                <FaChevronRight className="transition-transform duration-300 group-open:rotate-90 text-red-600" />
-              </summary>
-
-              {sidebarOpen && (
-                <ul className="mt-2 space-y-1 text-md bg-sidebar text-sidebar-foreground rounded-md p-2">
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                     <Link href="/auth/fazenda/cadastrar">Cadastrar Fazenda</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="/auth/fazenda/listar">Listar Fazendas</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="#">Atualizar Fazenda</Link>
-                  </li>
-                  <li className="rounded-md px-2 py-1 hover:bg-sidebar-accent hover:text-sidebar-accent-foreground">
-                    <Link href="#">Excluir Fazenda</Link>
-                  </li>
-                </ul>
-              )}
-            </details>
-          </nav>
-        </div>
-      </aside>
 
       {/* Conteúdo principal */}
       <main className="flex-1 p-10 transition-colors duration-500">
